+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+
+- Make the Video Controller optional 
+- added more documentation.
+
 ## 0.1.1
 
 - demo video added
@@ -7,7 +12,7 @@
 
 - Initial public release
 - Network video playback support
-- Reels mode with auto play/pause on visibility
+- Reels mode with autoplay/pause on visibility
 - Lifecycle aware video pause
 - Route aware auto pause
 - External controller support